move detected skill list to dedicated component
