+++ front/components/skills/ImportSkillsDialog.tsx
@@ -1,9 +1,5 @@
-import type { DetectedSkillSummary } from "@app/lib/skill_detection";
-import {
-  type DetectedSkillStatus,
-  isImportableSkillStatus,
-  parseGitHubRepoUrl,
-} from "@app/lib/skill_detection";
+import { DetectedSkillsList } from "@app/components/skills/DetectedSkillsList";
+import { isImportableSkillStatus, parseGitHubRepoUrl } from "@app/lib/skill_detection";
 import {
   useDetectSkillsFromFiles,
   useDetectSkillsFromRepo,
@@ -13,11 +9,8 @@ import {
 import type { LightWorkspaceType } from "@app/types/user";
 import {
   Button,
-  Checkbox,
-  Chip,
-  ContentMessage,
-  ContextItem,
   cn,
+  ContentMessage,
   Dialog,
   DialogContainer,
   DialogContent,
@@ -29,7 +22,6 @@ import {
   InformationCircleIcon,
   Input,
   PlusIcon,
-  PuzzleIcon,
   Spinner,
   Tabs,
   TabsContent,
@@ -49,15 +41,6 @@ interface ImportSkillsDialogProps {
   onClose: () => void;
   owner: LightWorkspaceType;
 }
-
-const STATUS_CHIP_LABEL: Record<
-  Exclude<DetectedSkillStatus, "ready">,
-  string
-> = {
-  name_conflict: "Skill name already in use",
-  skill_already_exists: "Override existing skill",
-  invalid: "Invalid skill format",
-};
 
 const ACCEPTED_FILE_TYPES = ".md,.zip";
 
@@ -277,75 +260,6 @@ export function ImportSkillsDialog({
         />
       </DialogContent>
     </Dialog>
-  );
-}
-
-interface DetectedSkillsListProps {
-  detectedSkills: DetectedSkillSummary[];
-  selectedNames: Set<string>;
-  toggleSkill: (name: string) => void;
-  isDetecting: boolean;
-  detectError: string | null;
-}
-
-function DetectedSkillsList({
-  detectedSkills,
-  selectedNames,
-  toggleSkill,
-  isDetecting,
-  detectError,
-}: DetectedSkillsListProps) {
-  return (
-    <>
-      {detectError && (
-        <ContentMessage
-          title="Detection failed"
-          icon={InformationCircleIcon}
-          variant="rose"
-          size="lg"
-        >
-          {detectError}
-        </ContentMessage>
-      )}
-      {isDetecting && (
-        <div className="flex items-center justify-center py-4">
-          <Spinner size="md" />
-        </div>
-      )}
-      {detectedSkills.length > 0 && (
-        <ContextItem.List>
-          {detectedSkills.map((skill) => (
-            <ContextItem
-              key={skill.name}
-              title={<span className="text-sm font-normal">{skill.name}</span>}
-              visual={
-                <div className="flex items-center gap-2">
-                  <Checkbox
-                    checked={selectedNames.has(skill.name)}
-                    disabled={!isImportableSkillStatus(skill.status)}
-                    onCheckedChange={() => toggleSkill(skill.name)}
-                  />
-                  <ContextItem.Visual visual={PuzzleIcon} />
-                </div>
-              }
-              action={
-                skill.status !== "ready" ? (
-                  <Chip
-                    label={STATUS_CHIP_LABEL[skill.status]}
-                    size="xs"
-                    color={
-                      skill.status === "skill_already_exists"
-                        ? "info"
-                        : "warning"
-                    }
-                  />
-                ) : undefined
-              }
-            />
-          ))}
-        </ContextItem.List>
-      )}
-    </>
   );
 }
